web/pip_requirements.txt
Flask-RESTful==0.3.8
marshmallow==3.5.2
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
gunicorn==20.0.4
Flask-Migrate==2.5.3
pytest-flask==1.0.0
boto3==1.13.19
dateparser==0.7.4
python-twitter==3.5
celery==4.4.4
